# mvp_sound_sentinel/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Sound Sentinel API",
    "description": "API для системы обнаружения звуков Sound Sentinel",
    "version": "1.0.0",
    "contact": {
      "name": "Sound Sentinel Team"
    }
  },
  "servers": [
    {
      "url": "https://localhost:8000",
      "description": "Development server"
    },
    {
      "url": "https://192.168.0.86:8000",
      "description": "Local network server"
    }
  ],
  "paths": {
...
    "/update_device/{device_id}": {
      "put": {
        "summary": "Обновить информацию об устройстве",
        "description": "Обновляет характеристики устройства (WiFi, CPU, температура и т.д.)",
        "tags": ["Devices"],
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "0d47631a-b266-47b7-9116-c551656974c5"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceUpdate"
              },
              "example": {
                "wifi_signal": 85,
                "microphone_info": "USB Microphone",
                "status": "online",
                "last_seen": "2026-04-19T23:46:00.000Z",
                "cpu_usage": 12.5,
                "device_temperature": 45.2
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Устройство обновлено",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessMessage"
                },
                "example": {
                  "message": "Device updated successfully"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeviceUpdate": {
        "type": "object",
        "properties": {
          "wifi_signal": {
            "type": "integer",
            "minimum": 0,
            "maximum": 100,
            "description": "Уровень WiFi сигнала в процентах",
            "example": 85
          },
          "microphone_info": {
            "type": "string",
            "description": "Информация о микрофоне",
            "example": "USB Microphone"
          },
          "status": {
            "type": "string",
            "enum": ["online", "offline"],
            "description": "Статус устройства",
            "example": "online"
          },
          "last_seen": {
            "type": "string",
            "format": "date-time",
            "description": "Время последнего подключения",
            "example": "2026-04-19T23:46:00.000Z"
          },
          "cpu_usage": {
            "type": "number",
            "description": "Использование CPU в процентах",
            "example": 12.5
          },
          "device_temperature": {
            "type": "number",
            "description": "Температура устройства в градусах Цельсия",
            "example": 45.2
          }
        }
      },
...
      "SuccessMessage": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Сообщение об успехе",
            "example": "Operation completed successfully"
          }
        }
      },
...
    }
  },
...
}